ai/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
piexif>=1.1.3
tqdm>=4.66.0
tensorflow>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0
